# development/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Synapse.API | openapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/work-orders": {
      "get": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示一覧を取得する。\nステータス・日付範囲でフィルタリング可能（製造指示一覧画面 SCR-MO-001 対応）。",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/WorkOrderStatus"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrderDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrderDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrderDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示を発行する（MO-001）。\n製造指示番号は MO-YYYYMMDD-NNNN 形式で自動採番される。\n発行者は JWT クレームから自動取得する。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": { }
              },
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}": {
      "get": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示を1件取得する。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示を更新する（MO-003）。\n完了・キャンセル済みは更新不可。着手中は数量変更不可。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}/cancel": {
      "post": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示をキャンセルする（MO-004）。発行済・着手中のみ可能。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}/start": {
      "post": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示を着手中に移行する（MO-007）。\n発行済（Issued）からのみ遷移可能。着手後は指示数量が変更不可になる。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-orders/{id}/complete": {
      "post": {
        "tags": [
          "WorkOrders"
        ],
        "summary": "製造指示を完了に移行する（MO-007）。\n着手中（InProgress）からのみ遷移可能。完了後は変更・キャンセル不可。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateWorkOrderRequest": {
        "required": [
          "productionPlanNumber",
          "itemId",
          "lotNumber",
          "quantity",
          "plannedStartDate",
          "plannedEndDate",
          "workInstruction"
        ],
        "type": "object",
        "properties": {
          "productionPlanNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "plannedStartDate": {
            "type": "string",
            "format": "date"
          },
          "plannedEndDate": {
            "type": "string",
            "format": "date"
          },
          "workInstruction": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "製造指示発行リクエスト。\nロット管理品目（Item.IsLotManaged=true）では LotNumber の指定を推奨する。"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateWorkOrderRequest": {
        "required": [
          "quantity",
          "plannedStartDate",
          "plannedEndDate",
          "workInstruction"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "plannedStartDate": {
            "type": "string",
            "format": "date"
          },
          "plannedEndDate": {
            "type": "string",
            "format": "date"
          },
          "workInstruction": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "製造指示更新リクエスト。\n着手中は Quantity の変更が拒否される（ドメインルール）。"
      },
...
      "WorkOrderDto": {
        "required": [
          "id",
          "workOrderNumber",
          "productionPlanNumber",
          "itemId",
          "itemCode",
          "itemName",
          "lotNumber",
          "quantity",
          "plannedStartDate",
          "plannedEndDate",
          "status",
          "workInstruction",
          "createdByUserId",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workOrderNumber": {
            "type": "string"
          },
          "productionPlanNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemCode": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "plannedStartDate": {
            "type": "string",
            "format": "date"
          },
          "plannedEndDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "$ref": "#/components/schemas/WorkOrderStatus"
          },
          "workInstruction": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WorkOrderStatus": {
        "type": "integer"
      }
    }
  },
...
}